feat: add CustomModalGroups component for group creation and integrate with academic schedule refactor: update CustomDataGrid and CustomDataGridColumns for improved data handling and actions
