+++ src/components/util/CustomDataGridColumns.tsx
@@ -19,7 +19,15 @@ export const CustomDataGridColumns = ({ headers, actions, checkbox }: CustomData
         <th key={header} className="border border-gray-300 px-4 py-2">
           {header}
         </th>
+        
       ))}
+       {
+        actions && (
+        <th className="px-4 py-2">
+          Acciones
+        </th>
+        )
+      }
     </tr>
     
   );
